==> SILVR_w_EDM/silvr_projects/4_geometry_stability.ipynb <==
# Running geometry stability metric
Notebook should be run as nohup.

This notebook can only be run after 3_analyse. Auto3D requires the smiles string of each molecule to generate a new conformation. This notebook is dependent on `data.csv` generated by 2_analysis. 

All non fragmented molecules (is_fragmented==False) are selected and AIMNet geometry optimisation applied. Output molecules are placed in the directory `geometry_stability/`, in subdirectories with the same name as from the experiments directory.

# Install Auto3D

#pip install Auto3D
#!conda install -c conda-forge torchani -y
#!conda install -c conda-forge ase -y#

# Current experiment is taking way too long to run. Redo without isomer enumeration on GPU=1

# Get SMILES file for experiment running

#TODO: need to retrive this from the last notebook
# SILVR + EDM, fragments 0072 + 0107
# experiment_names = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']

# BRIDGE + EDM, fragments 0072 + 0107
# experiment_names = ['exp_16','exp_17','exp_18','exp_19','exp_20','exp_21','exp_22']


# BRIDGE + EDM, fragments 0072 + 0107 (250 SAMPLES)
# experiment_names = ['exp_23','exp_24','exp_25','exp_26','exp_27','exp_28','exp_29']

# BRIDGE + EDM, fragments 0072 + 0107 (Explicit Hydrogens)
# experiment_names = ['exp_30','exp_31','exp_32','exp_33','exp_34','exp_35','exp_36']

# BRIDGE + EDM, NDM-1 fragments B + C (Explicit Hydrogens)
experiment_names = ['exp_37','exp_38','exp_39','exp_40','exp_41','exp_42','exp_43']


geometry_path = "geometry_stability/"
results_path = geometry_path+"results_fast/"
smiles_path = results_path+f"{experiment_names[0]}_to_{experiment_names[-1]}_smiles.smi"
# need to make this more streamlined...
data_csv = f"experiments/{experiment_names[0]}_to_{experiment_names[-1]}_data.csv"

import pandas as pd
import os

def create_new_id(mol_id, exp_id):
    """
    Auto3D can not take underscores, so making new unique molecule identifier with dash -
    """
    new_mol_id = "-".join(mol_id.split("_"))
    new_exp_id = "-".join(exp_id.split("_"))
    new_id = new_exp_id+"-"+new_mol_id
    
    return new_id

# Create SMILES file

if not os.path.exists(results_path):
    os.makedirs(results_path)
    
if not os.path.exists(smiles_path):
    print("Making new SMILES file")
    df = pd.read_csv(data_csv, usecols=["mol_id", "exp_id", "smiles", "is_fragmented"])
    df["new_id"] = df.apply(lambda x : create_new_id(x.mol_id,x.exp_id), axis=1)

    selection = df[df["is_fragmented"]==False][["smiles", "new_id"]]
    selection.to_csv(path_or_buf=smiles_path, sep=' ', header=False, index=False)
else:
    print("SMILES file already exists")

# Run Auto3D

from Auto3D.auto3D import options, main
import numpy as np

# this can take quite a while
if __name__ == "__main__":
    job_name = "results_fast"
    
    os.environ['MKL_THREADING_LAYER'] = 'GNU'#bothching
    
    args = options(smiles_path, k=1, enumerate_isomer=False, job_name=job_name, gpu_idx=1)  
    out = main(args)
