==> SILVR_w_EDM/silvr_projects/6_shapegauss.ipynb <==
<div class="alert alert-block alert-danger"> This notebook must be run in an OpenEye environment and requires and OpenEye licence file. </div>

# This notebook runs shapegauss scoring using openeye - this notebook must be run in the oepython environment. 

# SILVR + EDM, fragments 0072 + 0107
experiment_names = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']


# BRIDGE + EDM, fragments 0072 + 0107
# experiment_names = ['exp_16','exp_17','exp_18','exp_19','exp_20','exp_21','exp_22']

# BRIDGE + EDM, fragments 0072 + 0107 (250 SAMPLES)
# experiment_names = ['exp_23','exp_24','exp_25','exp_26','exp_27','exp_28','exp_29']

# BRIDGE + EDM, fragments 0072 + 0107 (Explicit Hydrogens)
# experiment_names = ['exp_30','exp_31','exp_32','exp_33','exp_34','exp_35','exp_36']

# BRIDGE + EDM, NDM-1 fragments B + C (Explicit Hydrogens)
# experiment_names = ['exp_37','exp_38','exp_39','exp_40','exp_41','exp_42','exp_43']

#TODO: Fix inconsistent naming of variables
exp_names = experiment_names
exp_list = [f"experiments/{x}" for x in exp_names]

import os
import glob
import openeye
import tqdm
# make sure to set the OE_LICENSE environment variable, the full path should be included, or else openeye will kill your kernel!
os.environ['OE_LICENSE'] = '/home/ian/oe_license.txt'
os.chmod('/home/ian/oe_license.txt', 0o755)

for exp in exp_list:
    exp_id = exp.split("/")[-1]
    
    if not os.path.exists(f'shapegauss/{exp_id}/'):
        os.makedirs(f'shapegauss/{exp_id}/')
    
    ligands = glob.glob(exp+"/*.txt")
    
    
    for ligand in ligands:
        mol_id = ligand.split("/")[-1].split(".")[0]

        with open(ligand, "r") as readfile:
            file = readfile.read()

        with open("tmp.xyz","w") as writefile:
            writefile.write(file)

        os.system(f'python docking_oe/RescorePoses.py -in tmp.xyz -out shapegauss/{exp_id}/{mol_id}.sdf -receptor docking_oe/mpro2.oedu -score Shapegauss')
