==> SILVR_w_EDM/silvr_projects/7_extract_shapegauss.ipynb <==
# Extract shapegauss results

This notebook takes `data_optim.csv` (outputted by 5_extract_geometry_stabilities). This notebook then creates the new csv `data_all.csv` which contains all the results from this project neccessary to create the metrics figure.

import pandas as pd  
import glob

# SILVR + EDM, fragments 0072 + 0107
# experiment_names = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']


# BRIDGE + EDM, fragments 0072 + 0107
# experiment_names = ['exp_16','exp_17','exp_18','exp_19','exp_20','exp_21','exp_22']
# BRIDGE + EDM, fragments 0072 + 0107 (250 SAMPLES)
# experiment_names = ['exp_23','exp_24','exp_25','exp_26','exp_27','exp_28','exp_29']

# BRIDGE + EDM, fragments 0072 + 0107 (Explicit Hydrogens)
experiment_names = ['exp_30','exp_31','exp_32','exp_33','exp_34','exp_35','exp_36']
#TODO: Fix inconsistent naming of variables
exp_names = experiment_names
shapegauss_exp_list = [f"shapegauss/{x}/" for x in exp_names]#shapegauss
experiment_path = "experiments/"

shapegauss_results = []
for path in shapegauss_exp_list:
    shapegauss_mol_list = glob.glob(path+"*.sdf")
    
    for mol_path in shapegauss_mol_list:
        mol_id = mol_path.split("/")[-1].split(".")[0]
        
        with open(mol_path, "r") as readfile:
            shapegauss_score = float(readfile.readlines()[-3])
            shapegauss_results.append([mol_id,shapegauss_score])

old_df = pd.read_csv(experiment_path+f"{experiment_names[0]}_to_{experiment_names[-1]}_data_optim.csv")
shapegauss_results_df = pd.DataFrame(shapegauss_results, columns=["mol_id", "shapegauss"]).set_index("mol_id")
df = pd.merge(old_df,shapegauss_results_df, on="mol_id",how='left')
df.to_csv(experiment_path+f"{experiment_names[0]}_to_{experiment_names[-1]}_data_all.csv",index=False)

shapegauss_results_df
