==> SILVR_w_EDM/silvr_projects/10c_extract_vina.ipynb <==
# Extract vina results

This notebook then creates the new csv `data_all.csv` which contains all the results from this project neccessary to create the metrics figure.

import pandas as pd  
import glob

# SILVR + EDM, fragments 0072 + 0107
# experiment_names = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']

# BRIDGE + EDM, fragments 0072 + 0107
# experiment_names = ['exp_16','exp_17','exp_18','exp_19','exp_20','exp_21','exp_22']
# BRIDGE + EDM, fragments 0072 + 0107 (250 SAMPLES)
# experiment_names = ['exp_23','exp_24','exp_25','exp_26','exp_27','exp_28','exp_29']

# BRIDGE + EDM, fragments 0072 + 0107 (Explicit Hydrogens)
experiment_names = ['exp_30','exp_31','exp_32','exp_33','exp_34','exp_35','exp_36']
#TODO: Fix inconsistent naming of variables
exp_names = experiment_names
vina_exp_list = [f"docking_data/{x}/" for x in exp_names]#vina without water
# vina_exp_list = [f"docking_data/{x}/" for x in exp_names] # TODO: vina with water
experiment_path = "experiments/"

# vina scores are already written to a csv file so there's no need to extract even further.
vina_df_list = [f"{x}_mpro_vina_docking_data.csv" for x in exp_names]
vina_results_df = pd.DataFrame()
# read in all the df from vina_df_list
for path in vina_df_list:
    vina_df = pd.read_csv(path)
    # get rid of the first row if the first column is not 'ligand_name'
    if vina_df.columns[0] != 'ligand_name':
        vina_df = vina_df.iloc[1:]
        #drop the error_message column
        vina_df = vina_df.drop(columns=['error_message'])
        #rename the columns
        vina_df = vina_df.rename(columns={'computed_dG':'error_message','protein_name':'computed_dG','ligand_name':'protein_name','Unnamed: 0':'ligand_name'})
    
    vina_df_subset = vina_df[['ligand_name', 'computed_dG']]
    df = pd.DataFrame(vina_df_subset)
    df['ligand_name'] = df['ligand_name'].str.replace('.sdf', '').str.replace('.pdbqt', '').str.replace('.pdb', '')
    vina_results_df = pd.concat([vina_results_df, df])
vina_results_df = vina_results_df.rename(columns={'ligand_name': 'mol_id', 'computed_dG': 'vina_score'})
vina_results_df

old_df = pd.read_csv(experiment_path+f"{experiment_names[0]}_to_{experiment_names[-1]}_data_all.csv")
# merge the two dataframes
df = pd.merge(old_df,vina_results_df, on="mol_id",how='left')
#!!! This over writes the results from earlier!!!! 
df.to_csv(experiment_path+f"{experiment_names[0]}_to_{experiment_names[-1]}_data_all.csv", index=False)
df
