==> SILVR_w_EDM/silvr_projects/12_dissertation_figure_prep.ipynb <==
import pandas as pd

# SILVR + EDM, fragments 0072 + 0107
experiment_names = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']

# BRIDGE + EDM, fragments 0072 + 0107
# experiment_names = ['exp_16','exp_17','exp_18','exp_19','exp_20','exp_21','exp_22']

# BRIDGE + EDM, fragments 0072 + 0107 (250 SAMPLES)
# experiment_names = ['exp_23','exp_24','exp_25','exp_26','exp_27','exp_28','exp_29']
# BRIDGE + EDM, fragments 0072 + 0107 (Explicit Hydrogens)
# experiment_names = ['exp_30','exp_31','exp_32','exp_33','exp_34','exp_35','exp_36']
df = pd.read_csv(f"experiments/{experiment_names[0]}_to_{experiment_names[-1]}_data_all.csv")
len(df)

'''
These are the experiments I'd like to visualise in pymol
SILVR - top-scoring molecule from r_s = 0, exp_9
top-scoring molecule from r_s = 2, exp_14

BRIDGE - top-scoring molecule from r_s = 0, exp_16
top-scoring molecule from r_s = 0.25 or  0.5 exp_18， exp_19
'''
experiments_to_visualise = ['exp_9','exp_14','exp_16','exp_18','exp_19']
#read in the pd dataframes for the experiments
df_list = [f"{x}_mpro_vina_docking_data.csv" for x in experiments_to_visualise]

top_scoring_molecules_df = pd.DataFrame()
for df_name in df_list:
    vina_results_df = pd.read_csv(f"{df_name}")
    if vina_results_df.columns[0] != 'ligand_name':
        vina_results_df = vina_results_df.iloc[1:]
        #drop the error_message column
        vina_results_df = vina_results_df.drop(columns=['error_message'])
        #rename the columns
        vina_results_df = vina_results_df.rename(columns={'computed_dG':'error_message','protein_name':'computed_dG','ligand_name':'protein_name','Unnamed: 0':'ligand_name'})

    # printout the top-scoring molecule for each experiment
    top_scoring_molecule = vina_results_df.sort_values(by='computed_dG', ascending=True).head(1)
    # top_scoring_molecule_df is a dataframe with 3 columns: exp_id, ligand_name and computed_dG
    top_scoring_molecules_df = pd.concat([top_scoring_molecules_df, top_scoring_molecule]) 
    
top_scoring_molecules_df




'''
These are the experiments I'd like to visualise in pymol
SILVR - top-scoring molecule from each SILVR rate, exp_9, exp_10, exp_11, exp_12, exp_13, exp_14, exp_15
'''
experiments_to_visualise = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']
#read in the pd dataframes for the experiments
df_list = [f"{x}_mpro_vina_docking_data.csv" for x in experiments_to_visualise]

top_scoring_molecules_df_vina = pd.DataFrame()
for df_name in df_list:
    vina_results_df = pd.read_csv(f"{df_name}")
    if vina_results_df.columns[0] != 'ligand_name':
        vina_results_df = vina_results_df.iloc[1:]
        #drop the error_message column
        vina_results_df = vina_results_df.drop(columns=['error_message'])
        #rename the columns
        vina_results_df = vina_results_df.rename(columns={'computed_dG':'error_message','protein_name':'computed_dG','ligand_name':'protein_name','Unnamed: 0':'ligand_name'})

    # printout the top-scoring molecule for each experiment
    top_scoring_molecule = vina_results_df.sort_values(by='computed_dG', ascending=True).head(1)
    # top_scoring_molecule_df is a dataframe with 3 columns: exp_id, ligand_name and computed_dG
    top_scoring_molecules_df_vina = pd.concat([top_scoring_molecules_df_vina, top_scoring_molecule]) 
    
top_scoring_molecules_df_vina

'''
These are the experiments I'd like to visualise in pymol
SILVR - top-scoring molecule from each SILVR rate, exp_9, exp_10, exp_11, exp_12, exp_13, exp_14, exp_15
'''
experiments_to_visualise = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']
#read in the pd dataframes for the experiments
df_list = [f"{x}_mpro_oe_docking_data.csv" for x in experiments_to_visualise]

oe_match_vina = pd.DataFrame()
for df_name in df_list:
    oe_results_df = pd.read_csv(f"{df_name}")
    if oe_results_df.columns[0] != 'ligand_name':
        oe_results_df = oe_results_df.iloc[1:]
        #drop the error_message column
        oe_results_df = oe_results_df.drop(columns=['error_message'])
        #rename the columns
        oe_results_df = oe_results_df.rename(columns={'computed_dG':'error_message','protein_name':'computed_dG','ligand_name':'protein_name','Unnamed: 0':'ligand_name'})
            # get rid of the _0 in the ligand_name column
        oe_results_df['ligand_name'] = oe_results_df['ligand_name'].str.replace(r'(_0)$', '',regex=True)

    # print out the matching ligands between top scoring molecules from vina and oe_results_df
    # oe_match = pd.merge(top_scoring_molecules_df_vina, oe_results_df, on='ligand_name')
    # oe_match_vina = pd.concat([oe_match_vina, oe_match])

    oe_match = oe_results_df[oe_results_df['ligand_name'].isin(top_scoring_molecules_df_vina['ligand_name'])]
    oe_match_vina = pd.concat([oe_match_vina, oe_match])

    
oe_match_vina

top_scoring_molecules_vina_list = top_scoring_molecules_df_vina['ligand_name'].to_list()
# check if vina top-scoring molecules are in the oe top-scoring molecules
top_scoring_molecules_df['ligand_name'].isin(top_scoring_molecules_vina_list)

import os
top_scoring_molecules_vina_list = top_scoring_molecules_df_vina['ligand_name'].to_list()
for i in top_scoring_molecules_vina_list:
    os.system(f'cp -r docking_results/diffusion_model/exp_9_to_36/vina_docking/with_water/unsorted_data/*{i}* dissertation_figures/vina')

'''
These are the experiments I'd like to visualise in pymol
SILVR - top-scoring molecule from each SILVR rate, exp_9, exp_10, exp_11, exp_12, exp_13, exp_14, exp_15
'''
experiments_to_visualise = ['exp_9','exp_10','exp_11','exp_12','exp_13','exp_14','exp_15']
#read in the pd dataframes for the experiments
df_list = [f"{x}_mpro_oe_docking_data.csv" for x in experiments_to_visualise]

top_scoring_molecules_df_oe = pd.DataFrame()
for df_name in df_list:
    oe_results_df = pd.read_csv(f"{df_name}")
    if oe_results_df.columns[0] != 'ligand_name':
        oe_results_df = oe_results_df.iloc[1:]
        #drop the error_message column
        oe_results_df = oe_results_df.drop(columns=['error_message'])
        #rename the columns
        oe_results_df = oe_results_df.rename(columns={'computed_dG':'error_message','protein_name':'computed_dG','ligand_name':'protein_name','Unnamed: 0':'ligand_name'})

    # printout the top-scoring molecule for each experiment
    top_scoring_molecule = oe_results_df.sort_values(by='computed_dG', ascending=True).head(1)
    # top_scoring_molecule_df is a dataframe with 3 columns: exp_id, ligand_name and computed_dG
    top_scoring_molecules_df_oe = pd.concat([top_scoring_molecules_df_oe, top_scoring_molecule]) 
    
top_scoring_molecules_df_oe

import os
top_scoring_molecules_oe_list = top_scoring_molecules_df_oe['ligand_name'].to_list()
for i in top_scoring_molecules_vina_list:
    os.system(f'cp -r docking_results/diffusion_model/exp_9_to_36/oe_docking/with_water/oe_docking_data/*{i}* dissertation_figures/oe')
